feat: refresh sidebar settings layout
